Renamed routing files. Not sure how I can make this much better... :(
