+++ core/route/allrouting.js
