+++ core/route/router.js
@@ -1,4 +1,4 @@
-const Routing = require('./routing');
+const Routing = require('./allrouting');
 const Express = require('express');
 /**
  * Router class, unlike the routes class this class
